+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 > 2015.12 (by Raymond Xie)
 
 * webqq 内部做了修改，导致登录及通讯失败，经过分析，予以修复。感谢网友 ScienJus 和 thegfw 提供的帮助。
+* DEBUG=1 qqbot, 启用调试模式，输出通讯数据
 
 > 2015.10 (by Raymond Xie)